Clean up memory data collection
